Define the project
This explains who we are and what we're doing. Takashi: You can probably accept this file change somehow. It's just a request right now, and I think you can either reject the changes, delete the branch, or "merge" it into the master branch, which should just make the main readme look just like this one. It's a branch, though, so if you don't like it, it'll just sort of hang here.