Object to do recursive reveal
The file contains code and annotations for an object that "recursively" reveals tiles in a way that minimizes the maximum amount of stack required. It's not faster than recursion (it might be slightly slower, depending on how the server does dynamic arrays), but it also greatly reduces the risk of a stack overflow error. I'm sorry; I've been sick all weekend.